Cambiso nuevos

+++ admin/intranet/deletenotam.php
@@ -14,7 +14,21 @@
 		}
 		
 		
+		$sql2 = "SELECT * FROM  notams where id=$id ";
+
+	if (!$result2 = $db->query($sql2)) {
+
+		die('There was an error running the query  [' . $db->error . ']');
+
+	}
+
+	while ($row2 = $result2->fetch_assoc()) {
 		
+		
+		$imagen = $row2["foto"];
+	}
+	
+	unlink('./uploadsnotam/' . $imagen);
 		
 			$sql1 = "delete from notams where id=$id";
